On MacOS, raise RuntimeError if osascript exist with nonzero code
Under this change, if `/usr/bin/osascript -l JavaScript ...` exits with
a nonzero code then a RuntimeError is raised (#56). This avoids
convert() failing silently when used in a script.

Test plan: The new unit test checks that a RuntimeError is raised, if it
is running on a MacOS system where Microsoft Word is not available.

This change needs to be manually tested on a MacOS system where Microsoft Word
is available. It should not have affected behaviour on Windows.