On MacOS, explicitly check for whether Microsoft Word is available.
Issue #56. Instead of raising a RuntimeError with the message from the
JXA, which is a little cryptic, instead check for Microsoft Word before
even trying to run the JXA, and raise an EnvironmentError if not.

Test plan: Updated unit test reflects this. Needs manual testing on a
system that has Microsoft Word.