Don't explicitly check for whether Microsoft Word is available
PR #57, Issue #56

This reverts commit acf494e59619d69706b73d984cd80d18f8cde26b.

Eager checking could be slow especially if convert() is being run on a
lot of different functions. Let's just rely on the JXA exiting with a
nonzero return code and giving the message "Application can't be found"
if and only if Microsoft Word cannot be launched, and raise an
EnvironmentError (or OSError) in that case. If JXA exits with a nonzero
return code with any other message, raise RuntimeError.

See discussion at https://github.com/AlJohri/docx2pdf/pull/57#issuecomment-1028868178