Release Windows docx file lock after premature script exit
This fixes an issue where if an exception occurs while saving the docx as a pdf (ex. if the pdf file already exists and is being used by another application), then the script will exit without closing the docx. This makes the docx uneditable until the Word process is terminated.